Make requests to next-elasticsearch.nlb.ft.com
As part of the DNS migration we've had to move to a seperate DNS zone (nlb.ft.com).

We tried setting up with what everyone else is using, octodns in Financial-Times/dns, but unfortunatly there is an issue with the health checks not working for our US Elasticsearch cluster if the request has a host header ending in a period.

I've worked with Nayana's team to configure a geo-loadbalanced address using CloudFormation instead.