This commit is used for a talking point - please don't merge
I am adding this code in a quick and nasty way so I can open a PR to discuss it with other engineers. I want to see if an option like this would be a good idea but the code itself is quick and nasty and should be improved before being used.