Revert the AWS signed fetch work
This isn't working when we pull it into next-article and it hasn't been
released yet. We don't understand this work enough to debug it and get
it working and it's blocking our n-logger work from being merged and
released. We'd like to revert everything.

As it appeared in two different PRs, I decided to revert it commit by
commit. The PRs are:

  - https://github.com/Financial-Times/n-es-client/pull/86
  - https://github.com/Financial-Times/n-es-client/pull/93

The commit hashes I reverted in order are:

  - 826ca12fc547e74f85cf15bad498a816b65f14fb
  - b848fa405d6fc063a12985ac8b687e40f929e11c
  - 5f6b1f53c669ffccca2c15eb2518f309b9c07047
  - ffef3f82047fca6f122a0a4ea72c123d5647c5cb
  - ce7e591102d7f155d80f436d6ec60e348d76bd08
  - 8af8c242fc4133ada4c2707e1f3e5ba2a944ae23
  - 2c49ac9f489c77062fb92856ddc6094be1efc2cb
  - 0e5ec8f02b90d2f2b93b0e68ce7a20997a186a3a
  - 2ed85457b29b0cd25cbef1d0d127bb564a496bea
  - 698302b6638afdf1af55a0ddaea66d22638ad7ff
  - 6fd9cd5e17b812be496a7148dfbbfb29b7073403
  - 6fd9cd5e17b812be496a7148dfbbfb29b7073403
  - 2f229b8bfe4f63d45ef125a7a5adc243498ae73c
  - 48f79ce46107ca3d4c0de5a03840d232d86ce659

I'll be installing this branch in next-article before merging so that I
can be sure that I didn't miss anything and that it's safe to install
the latest version of n-es-client in our apps.